requirements.txt
opencv-contrib-python-headless
numpy
streamlit
tensorflow-cpu==2.3.0
scikit-image==0.18.3
imutils